packages/webtap/pyproject.toml
[project]
name = "webtap-tool"
version = "0.9.1"
description = "Terminal-based web page inspector for AI debugging sessions"
readme = "README.md"
authors = [
    { name = "Fredrik Angelsen", email = "fredrikangelsen@gmail.com" }
]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]
dependencies = [
    "beautifulsoup4>=4.13.5",
    "cryptography>=45.0.6",
    "datamodel-code-generator>=0.35.0",
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "lxml>=6.0.1",
    "msgpack-python>=0.5.6",
    "platformdirs>=4.4.0",
    "protobuf>=6.32.0",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "replkit2[all]>=0.12.0",
    "requests>=2.32.4",
    "transitions>=0.9.3",
    "uvicorn>=0.35.0",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
]

[tool.hatch.build]
exclude = [
  "/experimental",
]

[tool.hatch.build.targets.wheel]
packages = ["src/webtap"]

[project.scripts]
webtap = "webtap:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
replkit2 = { path = "../../../replkit2", editable = true }

[tool.basedpyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
pythonPlatform = "Linux"
include = ["src/**/*.py"]

[tool.ruff]
line-length = 120
include = ["src/**/*.py"]
